2.0

+++ produtos/models.py
@@ -12,5 +12,5 @@ class Produto(models.Model):
         nome = models.CharField(max_length=100)
         descricao = models.TextField()
         imagem = models.ImageField(upload_to='imagens', blank=True)
-        valor = models.IntegerField()
+        valor = models.FloatField()
         fator = models.CharField(choices=FATOR_CHOICES, max_length=1)
